requirements.txt
discord.py
PyNaCl
davey
loguru
python-dotenv
supertonic
google-cloud-texttospeech
aiosqlite
quart
